[#9] Use expiringdict to cache objects for v1
Replace the Server's global_cache with an instance of ExpiringDict
It will hold 100 items for upto 10 seconds. This will give us generally
up to date data when changes are made in V1's web interface.

A Version bump was required to v1pysdk to prevent a race condition

ExpiringDict requires OrderedDict which exists only in py27, so tox will
install the backport ordereddict from pypi when testing on py26